+++ CHANGELOG.md
@@ -3,3 +3,11 @@
 ## Fixed
 - Added Custom Edit Options
 - Update README
+
+## 0.1.3
+
+## Added
+- Can now add custom images in toast
+- Can Align Toast to top/bottom
+- Update README
+- Update Example